Fix issues of join operator
